requirements.txt
sentence-transformers
torch
PyMuPDF
faiss-cpu
transformers
numpy
google-generativeai
langchain_google_vertexai
python-dotenv
fastapi
uvicorn
sentence-transformers
slowapi
python-magic
langchain
langchain_community
langchain_core
langchain_huggingface
pydantic
pytz
Requests
chromadb
wikipedia
langchain_ollama
streamlit
langchain-chroma
pypdf

